Create KeyPadComponent.js
Used to display our keypad.Our keypad includes buttons which will do action when clicked. Since we need to send the click event to the parent ( with information about which button was clicked), we will call "this.props.OnClick" function on every button click and pass "e.target.name" as an argument.